# swagger.yml
openapi: "3.0.2"
info:
  title: API Title
  version: "1.0"
servers:
  - url: http://127.0.0.1:8000/
paths:
  /api/events/:
    get:
      description: "get news list"
      parameters:
        - name: page
          in: query
          description: page
          required: false
          schema:
            type: number
            default: 1
        - name: search
          in: query
          description: search text
          required: false
          schema:
            type: string
            default: ""
            example: "title"
          style: deepObject
          explode: true
        # - name: limit
        #   in: query
        #   required: false
        #   schema:
        #     type: number
        #     default: 20
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: number
                    example: 4
                    description: "Кол-во обьектов"
                  next:
                    type: string
                    example: "http://127.0.0.1:8000/api/events/?page=3"
                  previous:
                    type: string
                    example: "http://127.0.0.1:8000/api/events/?page=1"
                  results:
                    $ref: "#/components/schemas/Events"

  /api/event/<id>:
    get:
      description: "get event by id"
      parameters:
        - name: id
          in: path
          description: product id
          required: true
          schema:
            type: number
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventItemFull"

  /users/registration/:
    post:
      description: "registration"
      parameters:
        - name: username
          in: query
          description: username
          required: true
          schema:
            type: string
        - name: password
          in: query
          description: password
          required: true
          schema:
            type: string
      responses:
        "200":
          description: successful operation
        "500":
          description: internal server error

  /api/token/:
    post:
      description: "authorization"
      parameters:
        - name: username
          in: query
          description: username
          required: true
          schema:
            type: string
        - name: password
          in: query
          description: password
          required: true
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  refresh:
                    type: string
                    example: "token"
                  access:
                    type: string
                    example: "token"
        "401":
          description: No active account found with the given credentials

  /api/token/refresh/:
    post:
      description: "refresh token"
      parameters:
        - name: token
          in: query
          description: token
          required: true
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  refresh:
                    type: string
                    example: "token"
                  access:
                    type: string
                    example: "token"
        "401":
          description: Token is invalid or expired
components:
  schemas:
    Events:
      type: array
      items:
        $ref: "#/components/schemas/EventItem"

    EventItem:
      type: object
      properties:
        id:
          type: number
          example: 123
        title:
          type: string
          example: "name of event"
        participant_count:
          type: number
          example: 123
        city:
          type: string
          example: "Tula"
        event_time_start:
          type: string
          example: "2023-10-26 18:25"
        event_time_end:
          type: string
          example: "2023-10-26 18:25"
        preview:
          type: object
          $ref: "#/components/schemas/Preview"

    EventItemFull:
      type: object
      properties:
        id:
          type: number
          example: 123
        title:
          type: string
          example: "name of event"
        text:
          type: string
          example: "bio of event"
        preview:
          type: object
          $ref: "#/components/schemas/Preview"
        manager:
          type: object
          $ref: "#/components/schemas/Manager"
        city:
          type: string
          example: "Tula"
        street:
          type: string
          example: "Komsomola"
        house:
          type: string
          example: "3"
        coords:
          type: string
          example: "3.4243243242, 5.4324234143"
        event_time_start:
          type: string
          example: "2023-10-26 18:25"
        event_time_end:
          type: string
          example: "2023-10-26 18:25"
        reg_time_end:
          type: string
          example: "2023-10-26 18:25"
        time_until_reg_end:
          type: number
          example: 1
        participant_count:
          type: number
          example: 123
        tags:
          type: array
          $ref: "#/components/schemas/Tag"

    Manager:
      type: object
      properties:
        id:
          type: number
          example: 123
        first_name:
          type: string
          example: "Elena"
        last_name:
          type: string
          example: "Arefeva"
        mail:
          type: string
          example: "Elena@mail.com"
        avatar:
          type: string
          example: "http://127.0.0.1:8000/media/events_preview/KRvpx_QTcCg.jpg"
        phone:
          type: string
          example: "+79235553424"

    Tag:
      type: object
      properties:
        id:
          type: number
          example: 123
        title:
          type: string
          example: "name of event"

    Preview:
      type: object
      properties:
        src:
          type: string
          example: "/media/events_preview/KRvpx_QTcCg.jpg"